[BOT]: extract cache utilities
