+++ 2006Scape Client/src/main/java/core/CacheUtils.java
@@ -1,0 +1,66 @@
+package core;
+
+import net.Signlink;
+import util.Decompressor;
+
+import java.io.File;
+import java.io.FileInputStream;
+
+/**
+ * Utility methods for cache file operations extracted from {@link Game}.
+ */
+public final class CacheUtils {
+
+    private CacheUtils() {}
+
+    public static String getFileNameWithoutExtension(String fileName) {
+        File tmpFile = new File(fileName);
+        tmpFile.getName();
+        int whereDot = tmpFile.getName().lastIndexOf('.');
+        if (0 < whereDot && whereDot <= tmpFile.getName().length() - 2) {
+            return tmpFile.getName().substring(0, whereDot);
+        }
+        return "";
+    }
+
+    public static String indexLocation(int cacheIndex, int index) {
+        return Signlink.findcachedir() + "index" + cacheIndex + "/" + (index != -1 ? index + ".gz" : "");
+    }
+
+    public static void repackCacheIndex(int cacheIndex, Decompressor[] decompressors) {
+        System.out.println("Started repacking index " + cacheIndex + ".");
+        int indexLength = new File(indexLocation(cacheIndex, -1)).listFiles().length;
+        File[] file = new File(indexLocation(cacheIndex, -1)).listFiles();
+        try {
+            for (int idx = 0; idx < indexLength; idx++) {
+                int fileIndex = Integer.parseInt(getFileNameWithoutExtension(file[idx].toString()));
+                byte[] data = fileToByteArray(cacheIndex, fileIndex);
+                if (data != null && data.length > 0) {
+                    decompressors[cacheIndex].writeEntry(data.length, data, fileIndex);
+                    System.out.println("Repacked " + fileIndex + ".");
+                } else {
+                    System.out.println("Unable to locate index " + fileIndex + ".");
+                }
+            }
+        } catch (Exception e) {
+            System.out.println("Error packing cache index " + cacheIndex + ".");
+        }
+        System.out.println("Finished repacking " + cacheIndex + ".");
+    }
+
+    public static byte[] fileToByteArray(int cacheIndex, int index) {
+        try {
+            if (indexLocation(cacheIndex, index).length() <= 0 || indexLocation(cacheIndex, index) == null) {
+                return null;
+            }
+            File file = new File(indexLocation(cacheIndex, index));
+            byte[] fileData = new byte[(int) file.length()];
+            FileInputStream fis = new FileInputStream(file);
+            fis.read(fileData);
+            fis.close();
+            return fileData;
+        } catch (Exception e) {
+            return null;
+        }
+    }
+}
